pyproject.toml
[tool.poetry]
name = "fastapi-admin"
version = "0.1.0"
description = "Fast Admin Dashboard based on fastapi and tortoise-orm and rest-admin."
authors = ["long2ice <long2ice@prismslight.com>"]

[tool.poetry.dependencies]
python = "^3.8"
tortoise-orm = {git = "https://github.com/long2ice/tortoise-orm.git", branch = "long2ice"}
aiomysql = "*"
python-dotenv = "*"
uvloop = "*"
python-rapidjson = "*"
fastapi = {version = "*", extras = ["all"]}
aiosqlite = "*"
passlib = "*"
bcrypt = "*"
pyjwt = "*"
xlsxwriter = "*"
colorama = "*"
prompt_toolkit = "*"

[tool.poetry.dev-dependencies]
taskipy = "*"
asynctest = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.taskipy.tasks]
export = "poetry export -f requirements.txt --without-hashes > requirements.txt"